fix/ render html of mutated state after mutation occur
