+++ src/store.js
@@ -6,12 +6,14 @@ Vue.use(Vuex)
 export const store = new Vuex.Store({
   namespaced: true,
   state: {
-    data: []
+    data: [
+      {type: '', confidence: 0, start_pos: 0, end_pos: 0, text: 0}
+    ]
   },
 
   mutations: {
     updateData(state, payload){  
-      state.data.push(payload)
+      state.data.type = payload
     }
   }
 })
